feat: implement 16-bit quantization
- Quantize position, velocity, and radius into 16-bit integers

+++ ParallelCollisionLab/Network/NetworkServer.h
@@ -30,7 +30,7 @@ namespace Network
     private:
         void RegisterOrRefreshClient(const sockaddr_in& ClientAddr);
         void RemoveClient(const sockaddr_in& ClientAddr);
-        void SendHandshakeResponse(const sockaddr_in& Target, uint16_t SphereCount, float BoxHalfSize);
+        void SendHandshakeResponse(const sockaddr_in& Target, uint32_t SphereCount, float BoxHalfSize);
 
     private:
         SOCKET                        ServerSocket         = INVALID_SOCKET;
